@@ -1,8 +1,6 @@
 import { Router } from '@angular/router';
 import { Component } from '@angular/core';
 
-import { ToastyConfig } from 'ng2-toasty';
-
 @Component({
   selector: 'app-root',
   templateUrl: './app.component.html',
@@ -10,12 +8,7 @@ import { ToastyConfig } from 'ng2-toasty';
 })
 export class AppComponent {
 
-  constructor(
-    private toastyConfig: ToastyConfig,
-    private router: Router
-  ) {
-    this.toastyConfig.theme = 'bootstrap';
-  }
+  constructor(private router: Router) { }
 
   exibindoNavbar() {
     return this.router.url !== '/login';
