@@ -1,8 +1,8 @@
 import { Injectable } from '@angular/core';
 
-import { AuthHttp } from 'angular2-jwt';
 import 'rxjs/add/operator/toPromise';
 
+import { MoneyHttp } from './../seguranca/money-http';
 import { environment } from './../../environments/environment';
 
 @Injectable()
@@ -10,13 +10,13 @@ export class CategoriaService {
 
   categoriasUrl: string;
 
-  constructor(private http: AuthHttp) {
+  constructor(private http: MoneyHttp) {
     this.categoriasUrl = `${environment.apiUrl}/categorias`;
   }
 
   listarTodas(): Promise<any> {
-    return this.http.get(this.categoriasUrl)
+    return this.http.get<any>(this.categoriasUrl)
       .toPromise()
-      .then(response => response.json().content);
+      .then(response => response.content);
   }
 }
