@@ -3,7 +3,7 @@ import { Injectable } from '@angular/core';
 import 'rxjs/add/operator/toPromise';
 import * as moment from 'moment';
 
-import { AuthHttp } from 'angular2-jwt';
+import { MoneyHttp } from './../seguranca/money-http';
 import { environment } from './../../environments/environment';
 
 @Injectable()
@@ -12,22 +12,21 @@ export class DashboardService {
   lancamentosUrl: string;
 
   constructor(
-    private http: AuthHttp
+    private http: MoneyHttp
   ) {
     this.lancamentosUrl = `${environment.apiUrl}/lancamentos`;
   }
 
   lancamentosPorCategoria(): Promise<Array<any>> {
-    return this.http.get(`${this.lancamentosUrl}/estatisticas/por-categoria`)
-      .toPromise()
-      .then(response => response.json());
+    return this.http.get<Array<any>>(`${this.lancamentosUrl}/estatisticas/por-categoria`)
+      .toPromise();
   }
 
   lancamentosPorDia(): Promise<Array<any>> {
-    return this.http.get(`${this.lancamentosUrl}/estatisticas/por-dia`)
+    return this.http.get<Array<any>>(`${this.lancamentosUrl}/estatisticas/por-dia`)
       .toPromise()
       .then(response => {
-        const dados = response.json();
+        const dados = response;
         this.converterStringsParaDatas(dados);
 
         return dados;
