@@ -1,10 +1,10 @@
-import { URLSearchParams } from '@angular/http';
+import { HttpParams } from '@angular/common/http';
 import { Injectable } from '@angular/core';
 
 import 'rxjs/add/operator/toPromise';
-import { AuthHttp } from 'angular2-jwt';
 import * as moment from 'moment';
 
+import { MoneyHttp } from './../seguranca/money-http';
 import { environment } from './../../environments/environment';
 import { Lancamento } from './../core/model';
 
@@ -21,7 +21,7 @@ export class LancamentoService {
 
   lancamentoURL: string;
 
-  constructor(private http: AuthHttp) {
+  constructor(private http: MoneyHttp) {
     this.lancamentoURL = `${environment.apiUrl}/lancamentos`;
   }
 
@@ -30,31 +30,33 @@ export class LancamentoService {
   }
 
   pesquisar(filtro: LancamentoFiltro): Promise<any> {
-    const params = new URLSearchParams();
-    params.set('page', filtro.pagina.toString());
-    params.set('size', filtro.itensPorPagina.toString());
+    let params = new HttpParams({
+      fromObject: {
+        page: filtro.pagina.toString(),
+        size: filtro.itensPorPagina.toString()
+      }
+    });
 
     if (filtro.descricao) {
-      params.set('descricao', filtro.descricao);
+      params = params.append('descricao', filtro.descricao);
     }
 
     if (filtro.dataVencimentoInicio) {
-      params.set('dataVencimentoDe', moment(filtro.dataVencimentoInicio).format('YYYY-MM-DD'));
+      params = params.append('dataVencimentoDe', moment(filtro.dataVencimentoInicio).format('YYYY-MM-DD'));
     }
 
     if (filtro.dataVencimentoFim) {
-      params.set('dataVencimentoAte', moment(filtro.dataVencimentoFim).format('YYYY-MM-DD'));
+      params = params.append('dataVencimentoAte', moment(filtro.dataVencimentoFim).format('YYYY-MM-DD'));
     }
 
-    return this.http.get(`${this.lancamentoURL}?resumo`, { search: params })
+    return this.http.get<any>(`${this.lancamentoURL}?resumo`, { params })
       .toPromise()
       .then(response => {
-        const responseJson = response.json();
-        const lancamentos = responseJson.content;
+        const lancamentos = response.content;
 
         const resultado = {
           lancamentos,
-          total: responseJson.totalElements
+          total: response.totalElements
         };
 
         return resultado;
@@ -68,16 +70,15 @@ export class LancamentoService {
   }
 
   adicionar(lancamento: Lancamento): Promise<Lancamento> {
-    return this.http.post(this.lancamentoURL, JSON.stringify(lancamento))
-      .toPromise()
-      .then(response => response.json());
+    return this.http.post<Lancamento>(this.lancamentoURL, lancamento)
+      .toPromise();
   }
 
   atualizar(lancamento: Lancamento): Promise<Lancamento> {
-    return this.http.put(`${this.lancamentoURL}/${lancamento.codigo}`, JSON.stringify(lancamento))
+    return this.http.put<Lancamento>(`${this.lancamentoURL}/${lancamento.codigo}`, lancamento)
       .toPromise()
       .then(response => {
-        const lancamentoAlterado = response.json() as Lancamento;
+        const lancamentoAlterado = response;
 
         this.converterStringsParaDatas([lancamentoAlterado]);
 
@@ -86,10 +87,10 @@ export class LancamentoService {
   }
 
   buscarPorCodigo(codigo: number) {
-    return this.http.get(`${this.lancamentoURL}/${codigo}`)
+    return this.http.get<Lancamento>(`${this.lancamentoURL}/${codigo}`)
       .toPromise()
       .then(response => {
-        const lancamento = response.json() as Lancamento;
+        const lancamento = response;
 
         this.converterStringsParaDatas([lancamento]);
 
