@@ -1,10 +1,10 @@
 import { Injectable } from '@angular/core';
-import { ResponseContentType, URLSearchParams } from '@angular/http';
+import { HttpParams } from '@angular/common/http';
 
 import 'rxjs/add/operator/toPromise';
-import { AuthHttp } from 'angular2-jwt';
 import * as moment from 'moment';
 
+import { MoneyHttp } from './../seguranca/money-http';
 import { environment } from './../../environments/environment';
 
 @Injectable()
@@ -12,17 +12,16 @@ export class RelatorioService {
 
   lancamentosUrl: string;
 
-  constructor(private http: AuthHttp) {
+  constructor(private http: MoneyHttp) {
     this.lancamentosUrl = `${environment.apiUrl}/lancamentos`;
   }
 
   relatorioLancamentosPorPessoa(inicio: Date, fim: Date) {
-    const params = new URLSearchParams();
-    params.set('inicio', moment(inicio).format('YYYY-MM-DD'));
-    params.set('fim', moment(fim).format('YYYY-MM-DD'));
+    const params = new HttpParams()
+      .append('inicio', moment(inicio).format('YYYY-MM-DD'))
+      .append('fim', moment(fim).format('YYYY-MM-DD'));
 
-    return this.http.get(`${this.lancamentosUrl}/relatorios/por-pessoa`, { search: params, responseType: ResponseContentType.Blob })
-      .toPromise()
-      .then(response => response.blob());
+    return this.http.get(`${this.lancamentosUrl}/relatorios/por-pessoa`, { params, responseType: 'blob' })
+      .toPromise();
   }
 }
