@@ -1,7 +1,7 @@
 import { Injectable } from '@angular/core';
-import { Http, Headers } from '@angular/http';
+import { HttpClient, HttpHeaders } from '@angular/common/http';
 
-import { JwtHelper } from 'angular2-jwt';
+import { JwtHelperService } from '@auth0/angular-jwt';
 import 'rxjs/add/operator/toPromise';
 
 import { environment } from './../../environments/environment';
@@ -13,29 +13,28 @@ export class AuthService {
   jwtPayload: any;
 
   constructor(
-    private http: Http,
-    private jwtHelper: JwtHelper
+    private http: HttpClient,
+    private jwtHelperService: JwtHelperService
   ) {
     this.oauthTokenUrl = `${environment.apiUrl}/oauth/token`;
     this.carregarToken();
   }
 
   login(usuario: string, senha: string): Promise<void> {
-    const headers = new Headers();
-    const body = `username=${usuario}&password=${senha}&grant_type=password`;
-    headers.append('Content-Type', 'application/x-www-form-urlencoded');
-    headers.append('Authorization', 'Basic YW5ndWxhcjpAbmd1bEByMA==');
+    const headers = new HttpHeaders()
+      .append('Content-Type', 'application/x-www-form-urlencoded')
+      .append('Authorization', 'Basic YW5ndWxhcjpAbmd1bEByMA==');
 
-    return this.http.post(this.oauthTokenUrl, body, { headers, withCredentials: true })
+    const body = `username=${usuario}&password=${senha}&grant_type=password`;
+
+    return this.http.post<any>(this.oauthTokenUrl, body, { headers, withCredentials: true })
       .toPromise()
       .then(response => {
-        this.armazenarToken(response.json().access_token);
+        this.armazenarToken(response.access_token);
       })
       .catch(response => {
         if (response.status === 400) {
-          const responseJson = response.json();
-
-          if (responseJson.error === 'invalid_grant') {
+          if (response.error === 'invalid_grant') {
             return Promise.reject('Usuário ou senha inválida!');
           }
         }
@@ -45,15 +44,16 @@ export class AuthService {
   }
 
   obterNovoAccessToken(): Promise<void> {
-    const headers = new Headers();
-    const body = 'grant_type=refresh_token';
-    headers.append('Content-Type', 'application/x-www-form-urlencoded');
-    headers.append('Authorization', 'Basic YW5ndWxhcjpAbmd1bEByMA==');
+    const headers = new HttpHeaders()
+      .append('Content-Type', 'application/x-www-form-urlencoded')
+      .append('Authorization', 'Basic YW5ndWxhcjpAbmd1bEByMA==');
 
-    return this.http.post(this.oauthTokenUrl, body, { headers, withCredentials: true })
+    const body = 'grant_type=refresh_token';
+
+    return this.http.post<any>(this.oauthTokenUrl, body, { headers, withCredentials: true })
       .toPromise()
       .then(response => {
-        this.armazenarToken(response.json().access_token);
+        this.armazenarToken(response.access_token);
 
         console.log('Novo access token criado!');
 
@@ -73,7 +73,7 @@ export class AuthService {
   isAccessTokenInvalido() {
     const token = localStorage.getItem('token');
 
-    return !token || this.jwtHelper.isTokenExpired(token);
+    return !token || this.jwtHelperService.isTokenExpired(token);
   }
 
   temPermissao(permissao: string) {
@@ -91,7 +91,7 @@ export class AuthService {
   }
 
   private armazenarToken(token: string) {
-    this.jwtPayload = this.jwtHelper.decodeToken(token);
+    this.jwtPayload = this.jwtHelperService.decodeToken(token);
     localStorage.setItem('token', token);
   }
 
